+++ src/trackers/core/bytetrack/utils.py
@@ -19,7 +19,7 @@ def _get_alive_tracklets(
 ) -> list[T_ByteTrackTracklet]:
     """
     Remove dead or immature lost tracklets and get alive trackers
-    that are within `maximum_frames_without_update` AND (it's mature OR
+    that are within `maximum_frames_without_update` missed frames AND (it's mature OR
     it was just updated).
 
     Note:
@@ -57,6 +57,6 @@ def _get_alive_tracklets(
             tracklet.number_of_successful_consecutive_updates >= minimum_consecutive_frames
         )
         is_active = tracklet.time_since_update == 0
-        if tracklet.time_since_update < maximum_frames_without_update and (is_mature or is_active):
+        if tracklet.time_since_update <= maximum_frames_without_update and (is_mature or is_active):
             alive_tracklets.append(tracklet)
     return alive_tracklets