+++ src/trackers/core/sort/utils.py
@@ -19,7 +19,7 @@ def _get_alive_tracklets(
 ) -> list[T_SORTTracklet]:
     """
     Remove dead or immature lost tracklets and get alive trackers
-    that are within `maximum_frames_without_update` AND (it's mature OR
+    that are within `maximum_frames_without_update` missed frames AND (it's mature OR
     it was just updated).
 
     Note:
@@ -42,6 +42,6 @@ def _get_alive_tracklets(
     for tracklet in tracklets:
         is_mature = tracklet.number_of_successful_updates >= minimum_consecutive_frames
         is_active = tracklet.time_since_update == 0
-        if tracklet.time_since_update < maximum_frames_without_update and (is_mature or is_active):
+        if tracklet.time_since_update <= maximum_frames_without_update and (is_mature or is_active):
             alive_tracklets.append(tracklet)
     return alive_tracklets